+++ src/main/java/com/solvd/tickets/TicketsApplication.java
@@ -3,10 +3,12 @@ package com.solvd.tickets;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
  * Application.
  */
+@EnableScheduling
 @SpringBootApplication
 @ConfigurationPropertiesScan
 public class TicketsApplication {
