--- src/main/java/com/solvd/tickets/config/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Configuration.
- */
-package com.solvd.tickets.config;
