+++ src/main/java/com/solvd/tickets/jira/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Jira.
+ * Jira Integration.
  */
 package com.solvd.tickets.jira;
