+++ src/main/java/com/solvd/tickets/model/Ticket.java
@@ -5,18 +5,18 @@ package com.solvd.tickets.model;
  */
 public interface Ticket {
 
-    /**
-     * Id.
-     *
-     * @return Id
-     */
-    String id();
+  /**
+   * Id.
+   *
+   * @return Id
+   */
+  String id();
 
-    /**
-     * Summary.
-     *
-     * @return Summary
-     */
-    String summary();
+  /**
+   * Summary.
+   *
+   * @return Summary
+   */
+  String summary();
 
 }
