+++ src/main/resources/application.yaml
@@ -1,3 +1,5 @@
+application:
+  delay: ${APP_DELAY}
 server:
   port: 8080
   shutdown: graceful
